--- src/components/Map/index.vue
@@ -1,21 +1,0 @@
-<template>
-    <div id='map'>
-        map
-    </div>
-</template>
-
-<script lang='ts' src='./index.ts'/>
-
-
-<!-- Add 'scoped' attribute to limit CSS to this component only -->
-<style scoped>
-html,
-body,
-#map {
-    height: 100%;
-}
-body {
-    margin: 0;
-    height: 100%;
-}
-</style>
