+++ src/components/MapView/Marker/DefaultSpotMarker.ts
