+++ src/components/MapView/index.ts
@@ -6,8 +6,8 @@ import L, { Marker } from 'leaflet';
 import { GeoJsonObject, GeometryObject, Feature, FeatureCollection } from 'geojson';
 import { findNearest, getDistance } from 'geolib';
 import { GeolibInputCoordinates } from 'geolib/es/types';
-import CurrentLocationMarker from '@/components/Map/Marker/CurrentLocationMarker';
-import DefaultSpotMarker from '@/components/Map/Marker/DefaultSpotMarker';
+import CurrentLocationMarker from '@/components/MapView/Marker/CurrentLocationMarker';
+import DefaultSpotMarker from '@/components/MapView/Marker/DefaultSpotMarker';
 import { MapViewGetters } from '@/store/modules/MapViewModule/MapViewGetters';
