+++ src/components/SpotList/index.ts
@@ -1,7 +1,7 @@
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import SpotItem from '@/components/SpotItem/index.vue';
 import { Spot, Coordinate } from '@/store/types';
-import { GeolocationWrapper } from '@/components/Map/GeolocationWrapper';
+import { GeolocationWrapper } from '@/components/MapView/GeolocationWrapper';
 import { LatLngExpression } from 'leaflet';
 import { getDistance } from 'geolib';
