+++ tests/unit/components/Map/mapDisplayPolygons.spec.ts
@@ -1,11 +1,11 @@
 import { mapViewGetters, mapViewMutations } from '@/store';
-import map from '@/components/Map/index.vue';
+import map from '@/components/MapView/index.vue';
 import { MapViewState } from '@/store/types';
 import { FeatureCollection } from 'geojson';
 import { shallowMount } from '@vue/test-utils';
 import { cloneDeep } from 'lodash';
 import { testMapViewState } from '../../../resources/testMapViewState';
-import { GeolocationWrapper } from '@/components/Map/GeolocationWrapper';
+import { GeolocationWrapper } from '@/components/MapView/GeolocationWrapper';
 
 const mapViewStateTestData: MapViewState = cloneDeep(testMapViewState);
