+++ tests/unit/components/Map/mapMarkerSelect.spec.ts
@@ -1,12 +1,12 @@
 import { mapViewGetters, mapViewMutations } from '@/store';
 import { MapViewState } from '@/store/types';
 import { shallowMount } from '@vue/test-utils';
-import { GeolocationWrapper } from '@/components/Map/GeolocationWrapper';
-import Map from '@/components/Map';
+import { GeolocationWrapper } from '@/components/MapView/GeolocationWrapper';
+import Map from '@/components/MapView';
 import 'leaflet/dist/leaflet.css';
 import { cloneDeep } from 'lodash';
 import { testMapViewState } from '../../../resources/testMapViewState';
-import DefaultSpotMarker from '@/components/Map/Marker/DefaultSpotMarker';
+import DefaultSpotMarker from '@/components/MapView/Marker/DefaultSpotMarker';
 
 const mapViewStoreTestData: MapViewState = cloneDeep(testMapViewState);
